pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "Steam App test api"
authors = ["Jojo Oatey <jojo.oatey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
sqlalchemy = "^1.4.45"
strawberry-graphql = {extras = ["debug-server"], version = "^0.151.0"}
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = "api.main:start"
